# C# decimal to Python conversion