--- a/Algorithm.Python/decimal.py
+++ b/Algorithm.Python/decimal.py
@@ -1,0 +1,18 @@
+# QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
+# Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+class Decimal(float):
+    '''This is required for backwards compatibility with users performing operations over expected decimal types.
+    NOTE: previously we converted C# decimal into python Decimal, but now its converted to python float due to performance.'''
+    def __init__(self, obj):
+        self = obj