<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- This PR is related to PR https://github.com/QuantConnect/pythonnet/pull/19 (but does not require it)
   > C# decimal will be cast to C# double and converted into python
   float
- Adding new `decimal.py` into the python algorithm project. This is
required for backwards compatibility with users performing operations over expected decimal types (like `Price`). `Decimal` class in `decimal.py` will cover the original `Decimal` with `float`.
- Updating two python regression test algorithms using python
execution models to be aware and ignore floating point precision errors
when handling order sizing.
- Adding new `NormalizeToStr()`. Required since now C# decimal will be converted to python float and trailing zeros were present after calling C# `Normalize()`

**Performance results with PythonNet of PR `QuantConnect/pythonnet/pull/19`**. ~15% improvement
>Master python_benchmark_algo
170.54 seconds at 42k data points per second. Processing total of 7,092,090 data points.
165.04 seconds at 43k data points per second. Processing total of 7,092,090 data points.
164.01 seconds at 43k data points per second. Processing total of 7,092,090 data points.
PR python_benchmark_algo
146.15 seconds at 49k data points per second. Processing total of 7,092,090 data points.
140.39 seconds at 51k data points per second. Processing total of 7,092,090 data points.
137.71 seconds at 51k data points per second. Processing total of 7,092,090 data points.

[python_benchmark_algo.txt](https://github.com/QuantConnect/Lean/files/2759935/python_benchmark_algo.txt)

>Master Py IndicatorRibbonBenchmark
339.04 seconds at 2k data points per second. Processing total of 782,223 data points.
335.39 seconds at 2k data points per second. Processing total of 782,223 data points.
307.85 seconds at 3k data points per second. Processing total of 782,223 data points.
PR Py IndicatorRibbonBenchmark -> **~45% improvement**
177.28 seconds at 4k data points per second. Processing total of 782,223 data points.
178.25 seconds at 4k data points per second. Processing total of 782,223 data points.
188.96 seconds at 4k data points per second. Processing total of 782,223 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Related to https://github.com/QuantConnect/Lean/issues/2825
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Python Execution Speed Improvement https://github.com/QuantConnect/Lean/projects/7
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->