@mchandschuh 2019-01-16
Any analysis done regarding potential loss of precision? Perhaps it would be most notable when running with raw prices.