@Martin-Molinero 2019-01-16
Floating point precision loss. I believe there are two aspects here: A- Python floats are mapped to IEEE-754 “double precision” which have 15 to 17 significant decimal digits which should be enough. B- Representation Error: `the fact that some (most, actually) decimal fractions cannot be represented exactly as binary (base 2) fractions`: ugly truth, but shouldn't have a realtive big impact either. Note that accounting will still be on the C# side calculated in decimals (interesting fact, and kind of reassuring, is that other platforms are also using double precision floating point).

ref https://docs.python.org/3/tutorial/floatingpoint.html https://en.wikipedia.org/wiki/Double-precision_floating-point_format